Make pango and fontconfig optional on *nix.
Unlike windows and GDI, there is no guarantee that these libraries will
exist on a *nix system (especially OS X).

However, the utilities provided by this library are useful even without
the font manipulation functions, so there's no reason the entire
library should require them to load.

If one of the libraries fails to load and the function that depends on
it is called, an error is thrown.